# Minor performance improvement