Have the processing thread get the entire buffer
available to reduce its mutex locks