@edwarnicke 2020-10-17
# Before
```
BenchmarkAll/Performance/ConcurrentAsyncMapAccess-2         	 1828819	       647 ns/op
BenchmarkAll/Performance/ConcurrentSyncMapAccessTest-2      	 1000000	      1096 ns/op
BenchmarkAll/Performance/ExecutorAsync-2                    	 3116842	       388 ns/op
BenchmarkAll/Performance/ExecutorSync-2                     	 1573186	       754 ns/op
BenchmarkAll/Tests/ASyncExec-2                              	 1285108	       927 ns/op
BenchmarkAll/Tests/SyncExec-2                               	 1472752	       809 ns/op
BenchmarkAll/Tests/ExecOrder1-2                             	 1000000	      1168 ns/op
BenchmarkAll/Tests/ExecOrder2-2                             	    6645	    261249 ns/op
BenchmarkAll/Tests/ExecOneAtATime-2                         	 2385934	       586 ns/op
BenchmarkAll/Tests/DataRace-2                               	 1256661	       956 ns/op
BenchmarkAll/Tests/NestedAsyncDeadlock-2                    	 4682691	       362 ns/op
```

#After

```
BenchmarkAll/Performance/ConcurrentAsyncMapAccess-2         	 2648149	       401 ns/op
BenchmarkAll/Performance/ConcurrentSyncMapAccessTest-2      	 1000000	      1088 ns/op
BenchmarkAll/Performance/ExecutorAsync-2                    	 3740234	       313 ns/op
BenchmarkAll/Performance/ExecutorSync-2                     	 1653890	       742 ns/op
BenchmarkAll/Tests/ASyncExec-2                              	 1324326	       905 ns/op
BenchmarkAll/Tests/SyncExec-2                               	 1504466	       775 ns/op
BenchmarkAll/Tests/ExecOrder1-2                             	 1000000	      1212 ns/op
BenchmarkAll/Tests/ExecOrder2-2                             	    6709	    252991 ns/op
BenchmarkAll/Tests/ExecOneAtATime-2                         	 2186032	       526 ns/op
BenchmarkAll/Tests/DataRace-2                               	 1323343	       907 ns/op
BenchmarkAll/Tests/NestedAsyncDeadlock-2                    	 5247820	       336 ns/op
```