@edwarnicke 2020-10-18
Interesting datapoint:
```
BenchmarkAll/Performance/ConcurrentAsyncMapAccess-2         	 3008982	       337 ns/op
BenchmarkAll/Performance/ConcurrentSyncMapAccessTest-2      	 1272764	       898 ns/op
BenchmarkAll/Baselines/BaselineConcurrentAccess-2           	 6537829	       206 ns/op
BenchmarkAll/Baselines/BaselineSyncMapReadWrite-2           	 1308152	       856 ns/op
```

BaselineConcurrentAccess is simply running through a map[int]bool in a single goroutine.  The fact that we are only 67% slower than that in ConcurrentAsyncMapAccess and %335 slower for ConcurrentSyncMapAccessTest is pretty good (given we are concurrency safe).

More interestingly compared to BaselineSyncMapReadWrite which uses sync.Map, we are 60% faster with ConcurrentAsyncMapAccess and only %5 slower with ConcurrentAsyncMapAccess.